requirements.txt
alembic==1.0.10
autoflake==1.3
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
flask-oidc==1.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
httplib2==0.13.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.12
MarkupSafe==1.1.1
oauth2client==4.1.3
okta==0.0.4
pyasn1==0.4.5
pyasn1-modules==0.2.5
pyflakes==2.1.1
python-dateutil==2.8.0
python-editor==1.0.4
python-slugify==3.0.2
requests==2.22.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.4
text-unidecode==1.2
urllib3==1.25.3
Werkzeug==0.15.4
